[fix] 상태바 색상 즉시 변경되지 않는 이슈 대응
